Fix implicit conversion warnings
These were caught as MSBuild warnings on Windows.
In most cases, acknowledge that the conversions may lose precision by doing a static cast.
In one case we change the type of an argument to be more accurate.

+++ orocos_kdl/src/jntarray.cpp
@@ -69,12 +69,12 @@ namespace KDL
 
     unsigned int JntArray::rows()const
     {
-        return data.rows();
+        return static_cast<unsigned int>(data.rows());
     }
 
     unsigned int JntArray::columns()const
     {
-        return data.cols();
+        return static_cast<unsigned int>(data.cols());
     }
 
     void Add(const JntArray& src1,const JntArray& src2,JntArray& dest)
