+++ orocos_kdl/src/treeiksolverpos_online.cpp
@@ -26,7 +26,7 @@
 
 namespace KDL {
 
-TreeIkSolverPos_Online::TreeIkSolverPos_Online(const double& nr_of_jnts,
+TreeIkSolverPos_Online::TreeIkSolverPos_Online(const unsigned int& nr_of_jnts,
                                                const std::vector<std::string>& endpoints,
                                                const JntArray& q_min,
                                                const JntArray& q_max,
